making re-installation of the config folder ask you if you'd like to override each individual file

+++ lib/cloud_crowd/command_line.rb
@@ -160,8 +160,13 @@ Options:
       exit(1)
     end
     
-    # Install a file and log the installation.
+    # Install a file and log the installation. If we're overwriting a file, 
+    # offer a chance to back out.
     def install_file(source, dest, is_dir=false)
+      if File.exists?(dest)
+        print "#{dest} already exists. Overwrite it? (yes/no) "
+        return unless ['y', 'yes', 'ok'].include? gets.chomp.downcase
+      end
       is_dir ? FileUtils.cp_r(source, dest) : FileUtils.cp(source, dest)
       puts "installed #{dest}"
     end
